Style: 모바일 뷰 이상 백그라운드 영역과 컨텐츠 영역 분리
