+++ src/components/layout/index.tsx
@@ -1,3 +1,4 @@
 export { default as Header } from './Header/Header';
 export { default as Layout } from './Layout/Layout';
 export { default as BottomNav } from './BottomNav/BottomNav';
+export { default as BackgroundArea } from './BackgroundArea';
